Support project based shell script provisioner.
More flexible shell based provision before vagrant up or vagrant provision
run specify with the project-dependencies environment variable the
provision shell scripts in the provision/packages folder to execute. If
this env is not defined than the full provision will be executed like
before that change. In combination with the vagrant-git tool it is now
possible that one vagrant project or base box can be used for different
projects. To get the most flexible solution with a small base box and
configurable provision script. The vagrant-git tool will use this new
feature soon and full automaticly.